Moves TxmanagerError to a separate file
