+++ consensus/service/src/tx_manager/error.rs
@@ -1,0 +1,46 @@
+use failure::Fail;
+use mc_consensus_enclave::Error as ConsensusEnclaveError;
+use mc_ledger_db::Error as LedgerDbError;
+use mc_transaction_core::{tx::TxHash, validation::TransactionValidationError};
+
+#[derive(Clone, Debug, Fail)]
+pub enum TxManagerError {
+    #[fail(display = "Enclave error: {}", _0)]
+    Enclave(ConsensusEnclaveError),
+
+    #[fail(display = "Transaction validation error: {}", _0)]
+    TransactionValidation(TransactionValidationError),
+
+    #[fail(display = "Tx already in cache")]
+    AlreadyInCache,
+
+    #[fail(display = "Tx(s) not in cache ({:?})", _0)]
+    NotInCache(Vec<TxHash>),
+
+    #[fail(display = "Ledger error: {}", _0)]
+    LedgerDb(LedgerDbError),
+}
+
+impl From<ConsensusEnclaveError> for TxManagerError {
+    fn from(err: ConsensusEnclaveError) -> Self {
+        if let ConsensusEnclaveError::MalformedTx(transaction_validation_error) = err {
+            Self::TransactionValidation(transaction_validation_error)
+        } else {
+            Self::Enclave(err)
+        }
+    }
+}
+
+impl From<TransactionValidationError> for TxManagerError {
+    fn from(err: TransactionValidationError) -> Self {
+        Self::TransactionValidation(err)
+    }
+}
+
+impl From<LedgerDbError> for TxManagerError {
+    fn from(err: LedgerDbError) -> Self {
+        Self::LedgerDb(err)
+    }
+}
+
+pub type TxManagerResult<T> = Result<T, TxManagerError>;
